# Make SimpleEdgeIter immutable