@sbromberger 2017-09-11
10% is 10%. I'll take it!